Split StreamProducer/Consumer source files
